Releasing 0.1.0. Genericized handling of CSM inputs.
Separated Actor logic into PlayerActor, which can handle inputs automatically and properly delegate them to the state machines.

PlayerActor now has a generic way of handling collisions. In the event that it touches a collider with the ITrigger interface, it will retrieve a string from that object and pass that to its state chain.