Stat system fully functional now. HUGE breaking changes.
For implementations, every Actor now needs an accompanying "Stat" component. This component can be subclassed with fields for any number of Stats. Roslyn analyzers generate the boilerplate necessary to allow States to make temporary changes to these stats.

Work needs to be done in order to keep a log of what objects access stats and what changes are being made.